Refactored participantList into a different file.
