+++ src/app/(pages)/pool/[pool-id]/participants/_components/participantCard.tsx
